Move template function into standalone file
